Release 2.0.0 (#7)
* Dev: Record Function Finished

* Dev: add progress bar when exporting videos

* Dev: add forgetting to terminate recording notification

* Dev-Patch: apply config settings to forgetting to terminate detector

* Dev: add forgetting to record notification

* Dev-Patch: Fix app quiting bug
Description: If error boxes has prompted, the application will not exit
when closing main window.
Reason: Electron will check whether all windows are closed(including
hidden windows), if the result is true, application will quit. But we
use Error Window Pool to reduce resources occupied. Electron will think
the unused error windows in the pool is alive and prevent quiting.

* Dev: add build support

* Dev-Patch: Add check for last video meta object

* Dev: Add Enter Binding of saving course name and abstract function

* Dev-UI: Modify the color of the progress bar of converting videos from blue to green because of cyy's vexatious requirement

* Dev-Patch: Fix the bug of not prompting forgetting to record
notification when first run

Reason: Because forgetToRecord.ts will check last video meta info to
confirm whether is recording. When the meta info is null, it will return
without checking.

When corgi-recorder first run, the meta info is null, and the check is
not carring on.

* Dev: Update license to GNU GPL v3

* Add CodeQL to analyze codes